+++ NEWS.md
@@ -4,7 +4,7 @@ This file documents the development of the package as well as open issues or poi
 
 
 ### Version in 3.0.0
-Date: 2021.03.07
+Date: 2021.03.16
 
 
 Changes:
@@ -12,11 +12,14 @@ Changes:
 *    Major version: Includes weighting functions to overcome biased norm samples
      Weights are built through the survey package. it is necessary to provide
      marginal means factor levels of stratification variables in the population
-*    weighted quantile estimation now draws on Harrell-Davis instead of Type7
-     by default
+*    new weighted quantile estimator: "inflation" (numerical solution for smaller
+     datasets)  
+*    weighted quantile estimation now draws on "inflation" and switches to Harrell-
+     Davis if too resource intense
 *    automatically remove  cases with missings in 'cnorm' function
 *    ppvt dataset exchanged with unstratified sample with additional background
      variables (migration, region, sex)
+